+++ client/web/src/nav/StatusMessagesNavItemQueries.ts
@@ -3,6 +3,12 @@ import { gql } from '@sourcegraph/http-client'
 export const STATUS_MESSAGES = gql`
     query StatusMessages {
         statusMessages {
+            ... on GitUpdatesDisabled {
+                __typename
+
+                message
+            }
+
             ... on CloningProgress {
                 __typename