+++ client/web/src/site-admin/SiteAdminAllUsersPage/UserManagement/queries.tsx
@@ -51,6 +51,7 @@ export const USERS_MANAGEMENT_USERS_LIST = gql`
                     createdAt
                     lastActiveAt
                     deletedAt
+                    locked
                 }
             }
         }