+++ enterprise/cmd/repo-updater/internal/authz/metrics.go
@@ -63,11 +63,11 @@ var (
 	}, []string{"type"})
 	metricsPermsConsecutiveSyncDelay = promauto.NewGaugeVec(prometheus.GaugeOpts{
 		Name: "src_repoupdater_perms_syncer_perms_consecutive_sync_delay",
-		Help: "The duration in minutes between last and current complete premissions sync.",
+		Help: "The duration in seconds between last and current complete premissions sync.",
 	}, []string{"type"})
 	metricsPermsFirstSyncDelay = promauto.NewGaugeVec(prometheus.GaugeOpts{
 		Name: "src_repoupdater_perms_syncer_perms_first_sync_delay",
-		Help: "The duration in minutes it took for first user/repo complete perms sync after creation",
+		Help: "The duration in seconds it took for first user/repo complete perms sync after creation",
 	}, []string{"type"})
 	metricsItemsSyncScheduled = promauto.NewGaugeVec(prometheus.GaugeOpts{
 		Name: "src_repoupdater_perms_syncer_items_sync_scheduled",